requirements.txt
customtkinter
Pillow
fastapi
uvicorn
pyyaml
pandas
pyqrcode
netifaces
pypng
